pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.2.2",
    "uvicorn[standard]>=0.20.0",
    "python-dotenv>=0.21.0",
]
